chore(react-apollo): restructure the application
 - restructure the application

[Maintains #166791179]
