+++ puzzled_api/apps/sudoku/admin.py
