+++ puzzled_api/apps/sudoku/schema.py
