+++ puzzled_api/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for puzzled project.
+Django settings for puzzled_api project.
 
 Generated by 'django-admin startproject' using Django 2.1.7.
 
@@ -42,8 +42,8 @@ INSTALLED_APPS = [
 
     # After the default packages
     'graphene_django',
-    'puzzled.apps.sudoku',
-    'puzzled.apps.puzzled_front',
+    'puzzled_api.apps.sudoku',
+    'puzzled_ui',
     'webpack_loader',
 ]
 
@@ -57,7 +57,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'puzzled.urls'
+ROOT_URLCONF = 'puzzled_api.urls'
 
 TEMPLATES = [
     {
@@ -75,7 +75,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'puzzled.wsgi.application'
+WSGI_APPLICATION = 'puzzled_api.wsgi.application'
 
 
 # Database
@@ -125,7 +125,7 @@ USE_TZ = True
 STATIC_URL = '/static/'
 
 GRAPHENE = {
-    'SCHEMA': 'puzzled.schema'
+    'SCHEMA': 'puzzled_api.schema'
 }
 
 WEBPACK_LOADER = {
