+++ puzzled_api/urls.py
@@ -1,4 +1,4 @@
-"""puzzled URL Configuration
+"""puzzled_api URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.1/topics/http/urls/
@@ -21,6 +21,6 @@ from .schema import schema
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('react/', include('puzzled.apps.puzzled_front.urls')),
+    path('react/', include('puzzled_ui.urls')),
     path('graphql/', csrf_exempt(GraphQLView.as_view(graphiql=True, schema=schema))),
 ]
