+++ puzzled_ui/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class PuzzledFrontConfig(AppConfig):
-    name = 'puzzled.apps.puzzled_front'
+    name = 'puzzled.apps.puzzled_ui'
